+++ instrumentation/net/http/otelhttp/test/transport_test.go
@@ -233,7 +233,8 @@ func TestWithHTTPTrace(t *testing.T) {
 	assert.Equal(t, "httptrace.GetConn", spans[0].Name())
 	assert.Equal(t, "test_span", spans[1].Name())
 	assert.Equal(t, "HTTP GET", spans[2].Name())
-	assert.NotEmpty(t, spans[1].Parent().SpanID())
+	assert.NotEmpty(t, spans[0].Parent().SpanID())
+	assert.NotEmpty(t, spans[2].Parent().SpanID())
 	assert.Equal(t, spans[2].SpanContext().SpanID(), spans[0].Parent().SpanID())
 	assert.Equal(t, spans[1].SpanContext().SpanID(), spans[2].Parent().SpanID())
 }